Delete incidents-when.html
Incident reports should be written regarding any incident, big or small. The reason we write these incidents is because we want to keep track of people who are causing issues. Even if it's a small incident that may not require anything done against them we still want to keep them for future reference if the incident keeps occuring over and over.